requirements.txt
streamlit
pandas
uuid
python-barcode